+++ src/elements/GeoSelect/GeoSelect.vue
@@ -30,6 +30,7 @@
         :is-empty="!value"
         :is-value-deletable="isValueDeletable"
         :disabled="disabled"
+        :variant="variant"
         @click="toggleSelect"
         @delete-value="deleteValue"
       >